% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_file.R
\name{create_data_file}
\alias{create_data_file}
\title{Save viable data}
\usage{
create_data_file(info)
}
\arguments{
\item{info}{a dataframe created based on the digital skills programme excel file}
}
\value{
data.csv files saved in folders with the slug name
}
\description{
Take information from the master spreadsheet with all future workshops, and extract the relevant information for GitHub pages
}
