% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_local_folders.R
\name{create_local_folders}
\alias{create_local_folders}
\title{Create local folders
create folders inside a local folder called "files", folder names are future workshop slugs.}
\usage{
create_local_folders(df)
}
\arguments{
\item{df}{a dataframe created based on the digital skills programme excel file}
}
\description{
Create local folders
create folders inside a local folder called "files", folder names are future workshop slugs.
}
