% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sharepoint_folder.R
\name{create_sharepoint_folder}
\alias{create_sharepoint_folder}
\title{Create Sharepoint site}
\usage{
create_sharepoint_folder(drv, info)
}
\arguments{
\item{drv}{the Sharepoint Drive to check}

\item{info}{information about the workshop taken from the holy excel sheet}
}
\description{
Takes information about the workshop. Checks if a sharepoint folder for a workshop slug already exists.
Creates a folder if it does not exist yet. Throws an error if sharepoint folders cannot be retrieved.
(within the current workflow, this will most likely be from the "Instructors" Sharepoint page)
}
