% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ws_channel.R
\name{create_ws_channel}
\alias{create_ws_channel}
\title{Create Microsoft Teams Channel}
\usage{
create_ws_channel(instr_team, slug)
}
\arguments{
\item{instr_team}{a teams Microsoft 365 object}

\item{slug}{slug of the workshop. Usually will be input through info (from the holy excel sheet). info$slug will have this parameter}
}
\description{
Takes information about the workshop. Checks if a channel for a workshop slug already exists.
Creates a channel if it does not exist yet. Throws an error if channels cannot be retrieved.
(within the current workflow, this will most likely be from the "Instructors" channel)
}
