% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_comms_doc.R
\name{fill_comms_doc}
\alias{fill_comms_doc}
\title{fill comms doc}
\usage{
fill_comms_doc(comms_doc_info, meta_fld)
}
\arguments{
\item{comms_doc_info}{information about the workshop to be inserted into the different documents
should include: startdate, enddate, title, address, venue, starttime, endtime, slug}

\item{meta_fld}{the URL of the meta folder}
}
\value{
no output, just saving files within the project directory
}
\description{
Take the template word documents and save them with specific information for the upcoming workshop in the corresponding Sharepoint channel
}
