% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_future_workshops.R
\name{get_future_workshops}
\alias{get_future_workshops}
\title{Get future workshops}
\usage{
get_future_workshops(df, future = T, include_location = F, token = "")
}
\arguments{
\item{df}{A dataframe based on the digital skills programme excel file}

\item{future}{A boolean (True/False) that indicates whether only future workshops should be returned}
}
\value{
a checked and formatted dataframe
}
\description{
Take information from the master spreadsheet with all future workshops, and generate the relevant information for GitHub pages.
This function checks whether necessary columns are present, and generates formatted columns that can be used
as input for websites and other communication material. It also drops empty lines.
}
