% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_fld.R
\name{get_meta_fld}
\alias{get_meta_fld}
\title{Get meta folder}
\usage{
get_meta_fld(slug)
}
\arguments{
\item{slug}{workshop slug}
}
\value{
a variable meta_fld that has the URL to the meta folder in it
}
\description{
Take the slug and return the corresponding URL where the meta files live for the workshop in question
}
