% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_planning_doc_info.R
\name{get_planning_doc_info}
\alias{get_planning_doc_info}
\title{Get information for planning document}
\usage{
get_planning_doc_info(info)
}
\arguments{
\item{info}{named vector with information about the workshop to be inserted into the different documents
should include: slug, startdate, newch, eventbrite, lead_instructor, instructor, helper}
}
\value{
planning doc information to be used in rendering the doc
}
\description{
Get specific information for a single workshop, represented as one line in the processed Holy Excel Sheet.
This represents information about one workshop. The information that is returned in a format that
can be input to is to set up the planning document with \code{render_planning_doc.R}.
}
