% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_drive.R
\name{read_from_drive}
\alias{read_from_drive}
\title{Read a file from the SharePoint drive into R}
\usage{
read_from_drive(year = "2023", drive = "instructors")
}
\arguments{
\item{year}{The year for which the Holy Excel Sheet will be read}

\item{drive}{name of the NLeSC sharepoint drive (default is 'instructors')}
}
\value{
an imported R object
}
\description{
This function opens the main excel sheet for the eScience Center
digital skills workshops (a.k.a. the "Holy Excel Sheet").
In the 'year' argument, indicate the year for which the file should be read.
If the file path no longer matches, e.g.,
\verb{General/Digital Skills Workshops 2023.xlsx}, please fix this information in
the csv file in \code{inst/extdata/rawinfo.csv}.
}
