% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_comms_doc.R
\name{render_comms_doc}
\alias{render_comms_doc}
\title{Render communications doc
Checks if the upcoming workshop is online or in person. Takes the appropriate template Rmd document from the template
repository. Renders a filled in document with specific information for the workshop.
Rendered docs will be saved in html and docx formats (default). Files will be saved in "files" directory of
working directory, within a directory that has the same name as the slug of the workshop.}
\usage{
render_comms_doc(info, outformat = "")
}
\arguments{
\item{outformat}{if left empty, both html and docx are printed. otherwise, choose one with "html", or "docx".}
}
\description{
Render communications doc
Checks if the upcoming workshop is online or in person. Takes the appropriate template Rmd document from the template
repository. Renders a filled in document with specific information for the workshop.
Rendered docs will be saved in html and docx formats (default). Files will be saved in "files" directory of
working directory, within a directory that has the same name as the slug of the workshop.
}
