% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_debrief_doc.R
\name{render_debrief_doc}
\alias{render_debrief_doc}
\title{Render debriefing doc}
\usage{
render_debrief_doc(debrief_doc_info)
}
\arguments{
\item{debrief_doc_info}{information about the workshop to be inserted into the debriefing document}
}
\value{
no output, just save the doc
}
\description{
Take the markdown file and render a word planning doc to be used in workshops
}
