% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_planning_doc.R
\name{render_plan}
\alias{render_plan}
\title{Render planning doc}
\usage{
render_plan(plan_doc_info)
}
\arguments{
\item{plan_doc_info}{information about the workshop to be inserted into the planning document}
}
\value{
no output, just save the doc
}
\description{
Take the markdown file and render a word planning doc to be used in workshops
}
