% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_post_sharepoint.R
\name{save_post_sharepoint}
\alias{save_post_sharepoint}
\title{Write message to Channel}
\usage{
save_post_sharepoint(info, alert = "Fenne Riemslagh")
}
\arguments{
\item{info}{a named vector with workshop information based on the digital skills programme excel file}

\item{alert}{any other people that need to be alerted}
}
\description{
Write message to Channel
}
