% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_post_sharepoint.R
\name{save_sharepoint}
\alias{save_sharepoint}
\title{Save to Sharepoint}
\usage{
save_sharepoint(ws_dat)
}
\arguments{
\item{ws_dat}{a dataframe created based on the digital skills programme excel file}
}
\value{
a dataframe that should be saved as data.csv in the workshop folder in SharePoint
}
\description{
Take information from the master spreadsheet with all future workshops, and extract the relevant information for GitHub pages
}
