% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_viable_data.R
\name{save_viable_data}
\alias{save_viable_data}
\title{Save viable data}
\usage{
save_viable_data(dat_struct)
}
\arguments{
\item{dat_struct}{a dataframe created based on the digital skills programme excel file}
}
\value{
data.csv files saved in folders with the slug name
}
\description{
Take information from the master spreadsheet with all future workshops, and extract the relevant information for GitHub pages
}
