% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_docs.R
\name{upload_docs}
\alias{upload_docs}
\title{Upload docs}
\usage{
upload_docs(
  instr_team,
  info,
  comms = TRUE,
  planning = TRUE,
  debriefing = TRUE,
  datafl = TRUE
)
}
\arguments{
\item{instr_team}{teams object obtained through running Microsoft365R::get_team("Instructors")}

\item{info}{information about the workshop for which the documents need to be uploaded}

\item{comms}{\code{TRUE} if the debriefing document should be uploaded (default)}

\item{planning}{\code{TRUE} if the planning document should be uploaded (default)}

\item{debriefing}{\code{TRUE} if the debriefing document should be uploaded (default)}

\item{datafl}{\code{TRUE} if the data.csv file should be uploaded (default)}
}
\description{
Upload docs
}
