% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verification.R
\name{verify_post_gfw}
\alias{verify_post_gfw}
\title{Verify that the dataset has been obtained after \code{get_future_workshops()}}
\usage{
verify_post_gfw(df)
}
\arguments{
\item{df}{dataframe with workshops}
}
\description{
Verify that the dataset has been obtained after \code{get_future_workshops()}
}
