% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop_activate.R
\name{workshop_activate}
\alias{workshop_activate}
\title{Choose workshops and activate them}
\usage{
workshop_activate(df)
}
\arguments{
\item{df}{data frame with future workshops (use \code{get_future_workshops()})}
}
\description{
Choose workshops and activate them
}
