% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_files.R
\name{create_files}
\alias{create_files}
\title{Create files}
\usage{
create_files(info)
}
\arguments{
\item{info}{named vector with workshop information}
}
\description{
Take the template word documents and save them with specific information
for the upcoming workshop.
}
