% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_files.R
\name{create_files}
\alias{create_files}
\title{Create files}
\usage{
create_files(info, folder = ".")
}
\arguments{
\item{info}{named vector with workshop information}

\item{folder}{location where files should be stored}
}
\description{
Take the template Rmd documents and save them with specific information
for the upcoming workshop. Write the workshop info to a data.csv file.
}
