% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_files.R
\name{create_files}
\alias{create_files}
\title{Create files}
\usage{
create_files(ws_dat, meta_fld)
}
\arguments{
\item{ws_dat}{information about the workshop to be inserted into the different documents}

\item{meta_fld}{the URL of the meta folder}
}
\value{
no output, just saving files within the project directory
}
\description{
Take the template word documents and save them with specific information for the upcoming workshop in the corresponding Sharepoint channel
}
