% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comms_doc_info.R
\name{get_comms_doc_info}
\alias{get_comms_doc_info}
\title{Get information for communications document}
\usage{
get_comms_doc_info(info)
}
\arguments{
\item{info}{information about the workshop to be inserted into the different documents
should include: startdate, enddate, title, address, venue, starttime, endtime, slug.
Taken from Holy Excel Sheet}

\item{meta_url}{the URL of the meta folder}
}
\value{
a list of workshop information, formatted correctly so that the communication document will render properly
}
\description{
Take the template word documents and save them with specific information for the upcoming workshop in the corresponding Sharepoint channel
}
