% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comms_doc_info.R
\name{get_comms_doc_info}
\alias{get_comms_doc_info}
\title{Get information for communications document}
\usage{
get_comms_doc_info(info)
}
\arguments{
\item{info}{information about the workshop to be inserted into the different documents
should include: startdate, enddate, title, address, venue, starttime, endtime, slug.
Taken from Holy Excel Sheet}

\item{meta_url}{the URL of the meta folder}
}
\value{
a list of workshop information, formatted correctly so that the communication document will render properly
}
\description{
Get specific information for a single workshop, represented as one line in the df_struct variable.
This represents information about one workshop. The information that is returned is needed
to set up the communication document, and is taken from the meta folder.
}
