% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_debrief_doc_info.R
\name{get_debrief_doc_info}
\alias{get_debrief_doc_info}
\title{Get debriefing doc info}
\usage{
get_debrief_doc_info(info)
}
\arguments{
\item{info}{information read from Holy Excel sheet: dataframe with
information about the workshop - used to set up the debriefing document}
}
\value{
debrief doc information to be used in rendering the doc
}
\description{
Get specific information for a workshop needed to set up the debriefing document.
The input data is read from Holy Excel sheet.
}
