% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_future_workshops.R
\name{get_future_workshops}
\alias{get_future_workshops}
\title{TESTING VERSION Get future workshops}
\usage{
get_future_workshops(excelfile)
}
\arguments{
\item{excelfile}{A dataframe created based on the digital skills programme excel file}
}
\value{
A dataframe that should be saved as data.csv in the workshop folder in SharePoint
}
\description{
TESTTEST Take information from the master spreadsheet with all future workshops, and extract the relevant information for GitHub pages
}
