% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_url.R
\name{get_meta_url}
\alias{get_meta_url}
\title{Get meta folder}
\usage{
get_meta_url(slug)
}
\arguments{
\item{slug}{workshop slug}
}
\value{
url to the folder containing workshop metadata
}
\description{
Take the slug and return the corresponding URL where the meta files live for
the workshop in question.
}
