% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plan_doc_info.R
\name{get_plan_doc_info}
\alias{get_plan_doc_info}
\title{Get planning doc info}
\usage{
get_plan_doc_info(ws_dat, meta_fld)
}
\arguments{
\item{ws_dat}{information about the workshop to be inserted into the planning document}

\item{meta_fld}{the URL of the meta folder}
}
\value{
planning doc information to be used in rendering the doc
}
\description{
Get the specific information for the upcoming workshop so it can be used in the planning doc
}
