% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plan_doc_info.R
\name{get_plan_doc_info}
\alias{get_plan_doc_info}
\title{Get information for planning document}
\usage{
get_plan_doc_info(info)
}
\arguments{
\item{ws_dat}{workshop data: information read from Holy Excel sheet: dataframe with
information about the workshop - used to set up the planning document}
}
\value{
planning doc information to be used in rendering the doc
}
\description{
Get specific information for a single workshop, represented as one line in the df_struct variable.
This represents information about one workshop. The information that is returned in a format that
can be input to is to set up the planning document with \code{render_planning_doc.R}.
Input to this function should include: startdate, enddate, title, address, venue, starttime, endtime,
slug for a single workshop.
This information is taken from the Holy Excel Sheet
}
