% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_drive.R
\name{read_from_drive}
\alias{read_from_drive}
\title{Read a file from the SharePoint drive into R}
\usage{
read_from_drive(
  path = "General/Digital Skills Workshops 2022.xlsx",
  drive = "https://nlesc.sharepoint.com/sites/instructors"
)
}
\arguments{
\item{path}{Path of the file inside the sharepoint drive}

\item{drive}{URL of the sharepoint drive}
}
\value{
an imported R object
}
\description{
This function combines the \code{Microsoft365R::get_sharepoint_site()} function
and the \verb{rio::import()`` functions to connect to a sharepoint drive and download a file, then open it in R. Because it uses }rio::import()`,
the function is able to open a large number of different filetypes.
}
\details{
The defaults are set to open the main excel sheet for the eScience Center
digital skills workshops (a.k.a. the "Holy Excel Sheet").
}
