% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_comms_doc.R
\name{render_comms_doc}
\alias{render_comms_doc}
\title{Render communications doc}
\usage{
render_comms_doc(
  info,
  outformat = "",
  template_url =
    "https://raw.githubusercontent.com/esciencecenter-digital-skills/template-docs-coordination/master/",
  folder = "."
)
}
\arguments{
\item{info}{a dataframe with information about the upcoming workshop}

\item{outformat}{if left empty, both html and docx are printed. otherwise, choose one with "html", or "docx".}

\item{template_url}{the place where the template for the communication document is saved in Rmd format}

\item{folder}{location for output document(s)}
}
\description{
Checks if the upcoming workshop is online or in person. Takes the appropriate template Rmd document from the template
repository. Renders a filled in document with specific information for the workshop.
Rendered docs will be saved in html and docx formats (default). Files will be saved in the working directory.
}
