% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_debrief_doc.R
\name{render_debrief_doc}
\alias{render_debrief_doc}
\title{Render debriefing doc}
\usage{
render_debrief_doc(
  info,
 
    template_url = "https://raw.githubusercontent.com/esciencecenter-digital-skills/template-docs-coordination/master/"
)
}
\arguments{
\item{info}{information about the workshop to be inserted into the debriefing document}

\item{template_url}{the place where the template for the document is saved in Rmd format}
}
\description{
Take the markdown file and render a word planning doc to be used in workshops
}
