% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_docs.R
\name{upload_docs}
\alias{upload_docs}
\title{Upload documents to sharepoint}
\usage{
upload_docs(info, folder = ".", drive = "instructors")
}
\arguments{
\item{info}{information about the workshop taken from the holy excel sheet}

\item{folder}{the directory of the files that need to be uploaded}

\item{drive}{the name of the Sharepoint drive where the workshop folder lives}
}
\description{
Upload documents to sharepoint
}
